requirements.txt
# pip install -r requirements.txt

argparse==1.2.1
requests==1.2.0
pyyaml==3.10

# Testing
pytest==2.3.4
flexmock==0.9.6